web-backend/requirements.txt
flask
flask_cors
kafka-python
requests
browser_cookie3